Created a Past Events page
Created a Past Events page. Currently, the page is a copy of the home page in order to perserve the format, the past events button on the home page was also activated, but currently is a dead link. Right now I am committing just to be able to return if I mess up some of the pastEvents page when deleting the unnecessary aspects of the home page.