Updated Temporary Past Events Page
Past events page is now filled only with the past events slide, temporarily filled with three events, but can be updated to fit a more common structure in the future. The past events button on the main page now links to the past events page and the home button on the past events page now links back to the home of the website.